backend/requirements.txt
fastapi
uvicorn
asyncpg
pgvector
sentence-transformers
pypdf
pillow
pytesseract
dateparser
openai-whisper
beautifulsoup4
requests
google-generativeai
httpx
python-multipart
numpy
pydantic
pydantic-settings
